Fixed import cycles and cleaned code.
