+++ src/components/projectdisplay.js
@@ -1,28 +1,96 @@
 import newElement from '../rendering/newelement';
 import listElements from '../rendering/listelements';
 import mainContainer from './maincontainer';
-// eslint-disable-next-line import/no-cycle
-import NewTask from '../pages/newtask';
-// eslint-disable-next-line import/no-cycle
-import editTask from '../pages/edittask';
-import Storage from '../classes/storage';
+import storageManager from '../classes/storagemanager';
 
 const projectDisplay = (project) => {
   const priorityText = ['Low', 'Normal', 'High'];
 
-  const removeTask = (taskIndex) => {
-    const projects = Storage.getProjects();
-    const tasks = project.getTasks();
-    tasks.splice(taskIndex, 1);
-    tasks.forEach(
-      (task, taskIndex) => {
-        task.setIndex(taskIndex);
+  const selectedStatus = (priority, check) => {
+    if (parseInt(priority, 10) === check) {
+      return ['selected', 'selected'];
+    }
+    return null;
+  };
+
+  const taskForm = (tasktitle, description, duedate, priority, save) => {
+    return listElements(
+      newElement('div', 'page contact d-flex flex-column'),
+      newElement('h1', 'contact-title', 'Contact'),
+      newElement('p', 'contact-text', 'New task'),
+      listElements(
+        newElement('form'),
+        listElements(
+          newElement('div', 'form-group'),
+          newElement('label', null, 'Title:', null, ['for', 'title_task']),
+          newElement('input', 'form-control', null, null, ['id', 'title_task'], ['placeholder', 'Title task'], ['value', tasktitle]),
+        ),
+        listElements(
+          newElement('div', 'form-group'),
+          newElement('label', null, 'Description:', null, ['for', 'description_task']),
+          newElement('input', 'form-control', null, null, ['id', 'description_task'], ['placeholder', 'A little description of the task'], ['value', description]),
+        ),
+        listElements(
+          newElement('div', 'form-group'),
+          newElement('label', null, 'Due date:', null, ['for', 'duedate_task']),
+          newElement('input', 'form-control', null, null, ['id', 'duedate_task'], ['type', 'date'], ['value', duedate]),
+        ),
+        listElements(
+          newElement('div', 'form-group'),
+          newElement('label', null, 'Priority:', null, ['for', 'priority_task']),
+          listElements(
+            newElement('select', 'form-control', null, null, ['id', 'priority_task']),
+            newElement('option', null, 'Low', null, ['value', '1'], selectedStatus(priority, 1)),
+            newElement('option', null, 'Normal', null, ['value', '2'], selectedStatus(priority, 2)),
+            newElement('option', null, 'High', null, ['value', '3'], selectedStatus(priority, 3)),
+          ),
+        ),
+        newElement(
+          'button',
+          'btn btn-primary',
+          'Save',
+          save,
+        ),
+      ),
+    );
+  };
+
+  const editTask = (project, task) => {
+    const edittask = taskForm(
+      task.getTitle(),
+      task.getDescription(),
+      task.getDueDate(),
+      task.getPriority(),
+      () => {
+        storageManager.editRenderTask(
+          project,
+          task,
+          document.getElementById('title_task').value,
+          document.getElementById('priority_task').value,
+          document.getElementById('description_task').value,
+          document.getElementById('duedate_task').value,
+        );
+        mainContainer.display(projectDisplay(project));
       },
     );
-    project.setTasks(tasks);
-    projects[project.getIndex()] = project;
-    Storage.setProjects(projects);
-    mainContainer.display(projectDisplay(project));
+    return edittask;
+  };
+
+  const NewTask = (project) => {
+    const newtask = taskForm(
+      '', '', '', 1,
+      () => {
+        storageManager.addRenderTask(
+          project,
+          document.getElementById('priority_task').value,
+          document.getElementById('description_task').value,
+          document.getElementById('duedate_task').value,
+        );
+        mainContainer.display(projectDisplay(project));
+      },
+    );
+
+    return newtask;
   };
 
   const tasksListHTML = [];
@@ -45,7 +113,15 @@ const projectDisplay = (project) => {
         ),
         listElements(
           newElement('td'),
-          newElement('button', 'btn btn-sm btn-danger', 'Remove', () => removeTask(task.getIndex())),
+          newElement(
+            'button',
+            'btn btn-sm btn-danger',
+            'Remove',
+            () => {
+              storageManager.removeTask(project, task);
+              mainContainer.display(projectDisplay(project));
+            },
+          ),
         ),
       ),
     );
